refactor: menu folder
