+++ src/modules/system/repositories/IMenusRepository.ts
@@ -5,9 +5,9 @@ import { Menu } from "../entities/Menu";
 interface IMenusRepository {
     create(data: ICreateMenusDTO): Promise<void>;
     list(activeOnly: boolean, adminOnly: boolean): Promise<Menu[]>;
-    findById(user_id: string, id: string, onlyTransaction: boolean): Promise<Menu>;
-    save(target: IUpdateMenusDTO): Promise<void>;
-    delete(target: Menu): Promise<void>;
+    findById(id: string): Promise<Menu>;
+    save(menu: IUpdateMenusDTO): Promise<void>;
+    delete(menu: Menu): Promise<void>;
 }
 
 export { IMenusRepository };
